Add prod webpack config
